Image Uploading and Thumbnails
Added basic image uploading to image profiles. If an image is invalid,
it won't be uploaded. If it is valid, it is uploaded to images/users/,
and the file name is hashed using the Hash::unique() function. Basic
restrictions on file types and sizes for the images.

Image thumbnails are generated using functions/thumbnail.php. Simply
specify the file, width, and height using GET requests. If a file
doesn't exist, or some other basic error happens, the default profile
picture will be the default.png file specified in images/users.